+++ HSAPI/src/main/kotlin/com/hobbystock/repositories/CategoryRepository.kt
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface CategoryRepository : JpaRepository<CategoryEntity, Long> {
+interface CategoryRepository : JpaRepository<CategoryEntity, Int> {
   fun findByName(name: String): CategoryEntity?
 
   // Paginated methods
